Create a Slack Adapter and Handler for requests
I create a SlackAdapter using the adapter pattern to be able to interact
with the slack api using the slack-ruby-client. Creating a wrapper
around the slack client made it easier for this system to call slack
client methods.

Also the handler class made it easier to process the requests received
from the slack API. Thus, having the functionality moved away from the
TaskController into it's own scope.

RELATED:
* https://refactoring.guru/design-patterns/adapter
* https://github.com/slack-ruby/slack-ruby-client
* https://github.com/slack-ruby/slack-ruby-client/blob/master/lib/slack/web/api/endpoints/chat.rb